v2 pipeline: DOUBLE timestamps, stage1 export, captions JSON, patch script
- milvus_memory_v2: fix time field DataType.FLOAT → DOUBLE (float32 caused
  128s rounding at unix timestamps, breaking time-window eval filtering)
- build_memory_v2: save stage1 filtered frames + CSV after SigLIP pass;
  export v1-compatible captions JSON with v2 extra columns after build;
  add --data_dir arg
- patch_timestamps.py: utility to fix existing DBs — recovers exact revisit
  timestamps via position-matching against stage1 CSV (0.000s error)
- pipeline shell scripts for v1 baseline (all seqs) and v2 (seq0, seq4)
- setup_remembr_v1.sh for isolated v1 conda env

+++ remembr/memory/milvus_memory_v2.py
@@ -50,7 +50,7 @@ class MilvusWrapperV2:
         schema.add_field("bge_embedding", DataType.FLOAT_VECTOR, dim=BGE_DIM)
         schema.add_field("position", DataType.FLOAT_VECTOR, dim=3)
         schema.add_field("theta", DataType.FLOAT)
-        schema.add_field("time", DataType.FLOAT)          # raw unix timestamp (scalar)
+        schema.add_field("time", DataType.DOUBLE)         # raw unix timestamp (scalar) — DOUBLE for full precision
         schema.add_field("caption", DataType.VARCHAR, max_length=3000)
         schema.add_field("image_path", DataType.VARCHAR, max_length=1000)
         schema.add_field("is_revisit", DataType.FLOAT)    # 0.0 = original, 1.0 = revisit
